requirements.txt
django-allauth==65.2.0
drf-spectacular==0.27.2
django-environ==0.11.2
django-cors-headers==4.6.0
dj-rest-auth==7.0.0
django-oauth-toolkit==3.0.1
djangorestframework-simplejwt==5.3.1
djangorestframework==3.15.2
drf-spectacular-sidecar==2024.11.1
django-storages==1.14.2
boto3==1.34.34
Pillow==10.2.0